+++ derive_builder/CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changed
+- The standard library `collections` crate was merged into `alloc`
+  ([PR](https://github.com/rust-lang/rust/pull/42648)).
+  Correspondingly when using this crate within a crate with `#![no_std]` you must
+  use `#![feature(alloc)] extern crate alloc` in your crate,
+  was `#![feature(collections)] extern crate collections`.
+
 ## [0.5.0] - 2017-06-30
 
 ### Changed